% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{clean.sp.labels}
\alias{clean.sp.labels}
\title{Clean Species Labels}
\usage{
clean.sp.labels(tree.dat)
}
\arguments{
\item{tree.dat}{\code{\link[=character]{character()}} File of tree data}
}
\value{

}
\description{
Clean Species Labels
}
\examples{
clean.sp.labels(tree.dat)

}
