% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{create.SORTIE.DBH.classes.Meas0}
\alias{create.SORTIE.DBH.classes.Meas0}
\title{Create SORTIE DBH classes Measurement Zero}
\usage{
create.SORTIE.DBH.classes.Meas0(
  sizeClasses,
  dbhclassSize,
  SORTIE.tree.dat,
  all.meas.plot.SORTIE,
  main.plot.phf
)
}
\arguments{
\item{sizeClasses}{\code{\link[=character]{character()}} Size classes}

\item{dbhclassSize}{\code{\link[=character]{character()}} DBH size classes}

\item{SORTIE.tree.dat}{\code{\link[=character]{character()}} SORTIE tree data}

\item{all.meas.plot.SORTIE}{\code{\link[=character]{character()}} All meas plot SORTIE}

\item{main.plot.phf}{\code{\link[=character]{character()}} Main plot phf}
}
\value{

}
\description{
Create SORTIE DBH classes Measurement Zero
}
\examples{
create.SORTIE.DBH.classes.Meas0(sizeClasses,dbhclassSize,SORTIE.tree.dat,all.meas.plot.SORTIE, main.plot.phf)

}
