% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{experiment_design}
\alias{experiment_design}
\title{Experiment Design}
\usage{
experiment_design(Level1, Level2, Replicates)
}
\arguments{
\item{Level1}{Level 1}

\item{Level2}{Level 2}

\item{Replicates}{Replicates}
}
\value{

}
\description{
Experiment Design
}
\examples{
experiment_design(Level1, Level2, Replicates)

}
