% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{ExtractFiles}
\alias{ExtractFiles}
\title{Extract output files}
\usage{
ExtractFiles(itype, exname, onename, extime)
}
\arguments{
\item{itype}{\code{\link[=double]{double()}} '1' = extract only the given file, otherwise extract all files in the directory}

\item{exname}{\code{\link[=character]{character()}} The directory that contains the tar file(s) to be extracted}

\item{onename}{\code{\link[=character]{character()}} The single tar file to extract (only used if itype = 1)}

\item{extime}{\code{\link[=double]{double()}} Extract time?}
}
\value{
A list of the extracted files.
}
\description{
\code{ExtractFiles()} extracts all files from any .gz.tar files in the directory
}
\examples{
ExtractFiles(itype,exname,onename,extime)

}
