% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateNewParameterFiles.R
\name{findFileLine}
\alias{findFileLine}
\title{Find the location in a SORTIE base parameter file from variable name}
\usage{
findFileLine(rf, itype, varname, vargroup, varmaster)
}
\arguments{
\item{rf}{\code{\link[=character]{character()}} Base XML parameter file to be modified}

\item{itype}{\code{\link[=integer]{integer()}} File type}

\item{varname}{\code{\link[=character]{character()}} variable codename}

\item{vargroup}{\code{\link[=character]{character()}} variable group name}

\item{varmaster}{\code{\link[=character]{character()}} variable master group}
}
\description{
\code{findFileLine()} finds the line within the base SORTIE parameter file that the new values (file or dataframe)
identifies is a variable to be modified. This file line applies only to the base parameter file and is
passed to \code{replaceInfo()}.
}
\examples{
samplebasexml[1:4]
VariableNames[1:4,]
gmf_time_new
findFileLine(rf=samplebasexml,itype=1,varname="timesteps")

}
