% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{import.psp}
\alias{import.psp}
\title{Import PSP data}
\usage{
import.psp(r.path, dat.type, tsas)
}
\arguments{
\item{r.path}{\code{\link[=character]{character()}} File path}

\item{dat.type}{\code{\link[=character]{character()}} Data type}

\item{tsas}{\code{\link[=character]{character()}} tsas}
}
\value{

}
\description{
Import PSP data
}
\examples{
import.psp(r.path, dat.type, tsas)

}
