% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{import.SORTIE.output}
\alias{import.SORTIE.output}
\title{Import SORTIE Output}
\usage{
import.SORTIE.output(yr, folder.path, fl.name)
}
\arguments{
\item{yr}{\code{\link[=double]{double()}} Year}

\item{folder.path}{\code{\link[=character]{character()}} Folder path}

\item{fl.name}{\code{\link[=character]{character()}} File name}
}
\value{

}
\description{
Import SORTIE Output
}
\examples{
import.SORTIE.output(yr, folder.path, fl.name)

}
