% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{import_Yrs_SORTIE}
\alias{import_Yrs_SORTIE}
\title{Import Selected SORTIE Years}
\usage{
import_Yrs_SORTIE(yrs, folder.path, fl.name)
}
\arguments{
\item{yrs}{\code{\link[=double]{double()}} Years}

\item{folder.path}{\code{\link[=character]{character()}} Folder Path}

\item{fl.name}{\code{\link[=character]{character()}} File name}
}
\value{

}
\description{
Import Selected SORTIE Years
}
\examples{
import_Yrs_SORTIE(yrs, folder.path, fl.name)

}
