% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{importSORTIE}
\alias{importSORTIE}
\title{Import SORTIE}
\usage{
importSORTIE(yrs, folder.path, fl.name, param_ver, plotID)
}
\arguments{
\item{yrs}{\code{\link[=double]{double()}} Years}

\item{folder.path}{\code{\link[=character]{character()}} Folder path}

\item{fl.name}{\code{\link[=character]{character()}} File name}

\item{param_ver}{\code{\link[=character]{character()}} Parameter file version}

\item{plotID}{\code{\link[=character]{character()}} Plot ID}
}
\value{

}
\description{
Import SORTIE
}
\examples{
importSORTIE(yrs, folder.path, fl.name, param_ver, plotID)

}
