% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{InitTreeList}
\alias{InitTreeList}
\title{Initiate tree list}
\usage{
InitTreeList(initname, numDigits = 0, diamBinMin, diamBinMax, diaminc)
}
\arguments{
\item{diaminc}{}
}
\value{

}
\description{
Initiate tree list
}
