% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeFiles.R
\name{makeFiles}
\alias{makeFiles}
\title{Make new parameter files}
\usage{
makeFiles(base_path, param_path, xmls_path)
}
\arguments{
\item{base_path}{\code{\link[=character]{character()}} the file path to the base parameter file(s) location}

\item{param_path}{\code{\link[=character]{character()}} the file path to the parameter value file(s) location}

\item{xmls_path}{\code{\link[=character]{character()}} the file path to the new output parameter file(s) location}
}
\value{

}
\description{
\code{makeFiles()} takes a base parameter file and replaces specific
parameter values to generate new parameter files to be run in SORTIE. Common use of
this function would be to change the number of timesteps or initiate different starting stands
to create a series of parameter files for an experiment

Prior to running \code{\link[=makeFiles]{makeFiles()}}, you must run \code{\link[=prepInputs]{prepInputs()}}
to generate the parameter value files needed for updating.
}
\examples{
makeFiles(base_path, param_path, xmls_path)

}
