% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{old.create.SORTIE.DBH.classes}
\alias{old.create.SORTIE.DBH.classes}
\title{Create SORTIE DBH classes (old version)}
\usage{
old.create.SORTIE.DBH.classes(
  tree.dat,
  num.meas,
  sizeClasses,
  SORTIE.tree.dat,
  plot.SORTIE
)
}
\arguments{
\item{tree.dat}{\code{\link[=character]{character()}} Tree data}

\item{num.meas}{\code{\link[=double]{double()}} Number of measurements}

\item{sizeClasses}{\code{\link[=character]{character()}} Size classes}

\item{SORTIE.tree.dat}{\code{\link[=character]{character()}} SORTIE tree data}

\item{plot.SORTIE}{\code{\link[=character]{character()}} SORTIE plot}
}
\value{

}
\description{
Create SORTIE DBH classes (old version)
}
\examples{
old.create.SORTIE.DBH.classes(tree.dat,num.meas,sizeClasses,SORTIE.tree.dat,plot.SORTIE)

}
