% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParseXML.R
\name{ParseMap}
\alias{ParseMap}
\title{Parse Map}
\usage{
ParseMap(grid_data)
}
\arguments{
\item{grid_data}{\code{\link[=character]{character()}} Grid data}
}
\value{

}
\description{
Parse Map
}
\examples{
ParseMap(grid_data)

}
