% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DevelopmentFunctions.R
\name{parseMapOld}
\alias{parseMapOld}
\title{Parse spatial files}
\usage{
parseMapOld(xmlname)
}
\arguments{
\item{xmlname}{\code{\link[=character]{character()}} XML file name}
}
\value{
The output contained in grids from SORTIE run as a data frame
}
\description{
\code{parseMapOld()} takes the output XML file name and returns the outputs stored in maps.
}
\examples{
parseMap(grid_data)

}
