% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParseXML.R
\name{ParseXML}
\alias{ParseXML}
\title{Parse XML}
\usage{
ParseXML(xmlname)
}
\arguments{
\item{xmlname}{\code{\link[=character]{character()}} XML file name}
}
\value{

}
\description{
Parse XML
}
\examples{
ParseXML(xmlname)

}
