% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{PrepareFile}
\alias{PrepareFile}
\title{PrepareFile}
\usage{
PrepareFile(pfname)
}
\arguments{
\item{pfname}{\code{\link[=character]{character()}} File path and name of the parameter file with new values}
}
\value{
pf1
}
\description{
PrepareFile
}
\examples{
PrepareFile(paramFile)

}
