% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{PrepInputs}
\alias{PrepInputs}
\title{ExtractFiles}
\usage{
PrepInputs(inputfiles)
}
\arguments{
\item{inputfiles}{\code{\link[=character]{character()}} The textfile containing the files to update and updating}
}
\value{

}
\description{
ExtractFiles
}
\examples{
PrepInputs(inputfiles)

}
