% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{psp.meas}
\alias{psp.meas}
\title{PSP measure}
\usage{
psp.meas(tree.dat, study.plots)
}
\arguments{
\item{tree.dat}{\code{\link[=character]{character()}} Tree data}

\item{study.plots}{\code{\link[=character]{character()}} Study plots}
}
\value{

}
\description{
PSP measure
}
\examples{
psp.meas(tree.dat, study.plots)

}
