% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{psp.years.age}
\alias{psp.years.age}
\title{PSP years and age}
\usage{
psp.years.age(plot.SORTIE, tree.dat, samples.dt, age.crit)
}
\arguments{
\item{plot.SORTIE}{\code{\link[=character]{character()}} SORTIE plot}

\item{tree.dat}{\code{\link[=character]{character()}} Tree data}

\item{samples.dt}{\code{\link[=character]{character()}} Samples}

\item{age.crit}{\code{\link[=character]{character()}} Age criteria}
}
\value{

}
\description{
PSP years and age
}
\examples{
psp.years.age(plot.SORTIE, tree.dat, samples.dt, age.crit)

}
