% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{ReadPlotFile}
\alias{ReadPlotFile}
\title{Read plot file}
\usage{
ReadPlotFile(outdir)
}
\arguments{
\item{outdir}{\code{\link[=character]{character()}} Output directory}
}
\value{

}
\description{
\code{ReadPlotFile()} reads .out output files.
}
\examples{
ReadPlotFile(outdir)

}
