% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{RemoveRow}
\alias{RemoveRow}
\title{RemoveRow}
\usage{
RemoveRow(ln1, rf)
}
\arguments{
\item{ln1}{\code{\link[=double]{double()}} Row number to remove}

\item{rf}{\code{\link[=character]{character()}} Base XML file to be modified}
}
\value{

}
\description{
RemoveRow
}
\examples{
RemoveRow(lnm,xml1)

}
