% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReplaceInfo.R
\name{ReplaceInfo}
\alias{ReplaceInfo}
\title{Replace Info}
\usage{
ReplaceInfo(rf, VariableNames, pf1, ncols, newname)
}
\arguments{
\item{rf}{\code{\link[=character]{character()}} Base XML parameter file to be modified}

\item{VariableNames}{\code{\link[=list]{list()}} Variable Names translation file}

\item{pf1}{\code{\link[=character]{character()}} Parameter file with new values}

\item{ncols}{\code{\link[=double]{double()}} Number of columns}

\item{newname}{\code{\link[=character]{character()}} New file name}
}
\value{

}
\description{
Replace Info
}
\examples{
xml2 <- ReplaceInfo(xml1, VariableNames, pf1, ncols, newname)

}
