% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReplaceFunctions.R
\name{replaceParameter}
\alias{replaceParameter}
\title{Replace a parameter}
\usage{
replaceParameter(ln1, rf, varvalue)
}
\arguments{
\item{ln1}{\code{\link[=double]{double()}} Line number of parameter to replace}

\item{rf}{\code{\link[=character]{character()}} Base XML parameter file to be modified}

\item{varvalue}{\code{\link[=character]{character()}} Output file name}
}
\description{
\code{replaceParameter()} replaces a parameter in the base file with a new value.
}
\examples{
replaceParameter(ln1, rf, fname)

}
