% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{RunSortie}
\alias{RunSortie}
\title{RunSortie}
\usage{
RunSortie(fname, sortie_loc)
}
\arguments{
\item{fname}{\code{\link[=character]{character()}} File path and name to be run}

\item{sortie_loc}{SORTIE location '0'}
}
\value{

}
\description{
RunSortie
}
\examples{
RunSortie(paste0(xmls_path,Xmls2Run[ix]),0)

}
