% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{RunSortiePar}
\alias{RunSortiePar}
\title{Run SORTIE in parallel}
\usage{
RunSortiePar(fname, numcores, sortie_loc)
}
\arguments{
\item{fname}{File path and parameter file names to run}

\item{numcores}{How many cores. Right now, the number of cores is the same as the number of files to run, but we will
update this to allow serial runs passed across cores}

\item{sortie_loc}{'0' defaults to Program Files (x86) for SORTIE location. If stored elsewhere, sortie_loc = location}
}
\value{

}
\description{
\code{RunSortiePar()} is a wrapper function that passes parameter files to process parallel SORTIE simulations
}
\examples{
RunSortiePar(parameterfile.xml,numcores = 2, 0)

}
