% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{sel.psp}
\alias{sel.psp}
\title{Select and clean PSP data}
\usage{
sel.psp(
  samples.dt,
  tree.dat,
  BECzone,
  BEClabel,
  SiteSeriesOfInterest,
  MinRemeasInterval
)
}
\arguments{
\item{samples.dt}{\code{\link[=character]{character()}} Samples}

\item{tree.dat}{\code{\link[=character]{character()}} Tree data}

\item{BECzone}{\code{\link[=character]{character()}} BEC zone}

\item{BEClabel}{\code{\link[=character]{character()}} BEC label}

\item{SiteSeriesOfInterest}{\code{\link[=character]{character()}} Site Series of Interest}

\item{MinRemeasInterval}{\code{\link[=character]{character()}} Minimum remeas interval}
}
\value{

}
\description{
Select and clean PSP data
}
\examples{
sel.psp(samples.dt,tree.dat,BECzone,BEClabel,SiteSeriesOfInterest,MinRemeasInterval)

}
