% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{SORTIE.tree.create}
\alias{SORTIE.tree.create}
\title{Create SORTIE tree population}
\usage{
SORTIE.tree.create(
  sizeClasses,
  SORTIE.species,
  Min.Adult.DBH,
  Max.Seedling.Hgt.m
)
}
\arguments{
\item{sizeClasses}{\code{\link[=character]{character()}} Size classes}

\item{SORTIE.species}{\code{\link[=character]{character()}} SORTIE species}

\item{Min.Adult.DBH}{\code{\link[=double]{double()}} Minimum Adult DBH}

\item{Max.Seedling.Hgt.m}{\code{\link[=double]{double()}} Maximum seedling height (m)}
}
\value{

}
\description{
Create SORTIE tree population
}
\examples{
SORTIE.tree.create(sizeClasses,SORTIE.species,Min.Adult.DBH,Max.Seedling.Hgt.m)

}
