% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReplaceFunctions.R
\name{updateNumYears}
\alias{updateNumYears}
\title{Update Number of Years (timesteps)}
\usage{
updateNumYears(xmls, num_years)

updateNumYears(xmls, num_years)
}
\arguments{
\item{xmls}{\code{\link[=character]{character()}} name of xml files (includes pathway)}

\item{num_years}{\code{\link[=numeric]{numeric()}} number of years to update parameter file}
}
\value{



}
\description{
\code{updateNumYears()} Updates the number of years for a simulation

\code{updateNumYears()} Update the parameter file with a new timestep
}
\examples{
updateNumYears(xmls, num_years)

updateNumYears(xmls,num_years)

}
