% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReplaceFunctions.R
\name{updateStemMap}
\alias{updateStemMap}
\title{Update the parameter file with a stemmap}
\usage{
updateStemMap(xmls, MapNames, TreatmentBase)
}
\arguments{
\item{xmls}{\code{\link[=character]{character()}} name(s) of xml files (includes pathway)}

\item{MapNames}{\code{\link[=character]{character()}} name(s) of the text files to add to the parameter file (includes pathway)}

\item{TreatmentBase}{\code{\link[=vector]{vector()}} names that appear in both XMLs and Map names to link
which stemmap with which parameter file}
}
\value{

}
\description{
\code{updateStemMap()} Update the parameter file with a stemmap.
}
\examples{
updateStemMap(xmls,MapNames)

}
