% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutputFunctions.R
\name{extractFiles}
\alias{extractFiles}
\title{Extract output files}
\usage{
extractFiles(itype, exname, onename, extime)
}
\arguments{
\item{itype}{\code{\link[=double]{double()}} '1' = extract only the given file, otherwise extract all files in the directory}

\item{exname}{\code{\link[=character]{character()}} The directory that contains the tar file(s) to be extracted}

\item{onename}{\code{\link[=character]{character()}} The single tar file to extract (only used if itype = 1)}

\item{extime}{\code{\link[=double]{double()}} Extract time?}
}
\value{
The routine returns a list of the extracted files.
}
\description{
\code{extractFiles()} reads the directory and extract all files from any .gz.tar files that are present.
It also assumes that all the files in a single .gz.tar file were in the same directory. It looks at the
first file in the .gz.tar, determines how many subdirectories are present, and then strips all those subdirectories.
The extracted files will be placed in a new extracted directory. If you want to put them
somewhere else, then change the variable extractDir
}
\examples{
extractFiles(itype,exname,onename,extime)

}
