% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{FindFileLine}
\alias{FindFileLine}
\title{Find a file line}
\usage{
FindFileLine(rf, itype, varname, vargroup, varmaster)
}
\arguments{
\item{rf}{\code{\link[=character]{character()}} Base XML parameter file to be modified}

\item{itype}{\code{\link[=integer]{integer()}} File type}

\item{varname}{\code{\link[=character]{character()}} variable codename}

\item{vargroup}{\code{\link[=character]{character()}} variable group name}

\item{varmaster}{\code{\link[=character]{character()}} variable master group}
}
\value{

}
\description{
\code{FindFileLine()} finds the line within the base parameter file that is to be modified.
}
\examples{
FindFileLine(rf,itype,codename, groupname, mastergroup)

}
