% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{InitTreeList}
\alias{InitTreeList}
\title{Define the names of tree list}
\usage{
InitTreeList(initname, numDigits = 0, diamBinMin, diamBinMax, diaminc)
}
\arguments{
\item{initname}{Naming convention for rows of tree sizes \code{\link[=character]{character()}}}

\item{numDigits}{Decimal places for tree diameter class size in new values and base parameter file}

\item{diamBinMin}{Minimum diameter class size \code{\link[=numeric]{numeric()}}}

\item{diamBinMax}{Maximum diameter class size (\link{numeric})}

\item{diaminc}{}
}
\value{

}
\description{
\code{InitTreeList()} adds tree list names and diameter size classes to the variable translation file
}
\details{
This function adds new rows to the variable translation default to capture specific naming conventions
for the tree size class by species (tree list) defined in the new values file and the base parameter file
}
