% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeFiles.R
\name{MakeFiles}
\alias{MakeFiles}
\title{FindFileLine}
\usage{
MakeFiles(VarTranslation, base_path, param_path, xmls_path)
}
\arguments{
\item{VarTranslation}{\code{\link[=data.frame]{data.frame()}} File type}

\item{base_path}{\code{\link[=character]{character()}} variable codename}

\item{param_path}{\code{\link[=character]{character()}} variable group name}

\item{xmls_path}{\code{\link[=character]{character()}} variable master group}
}
\value{

}
\description{
FindFileLine
}
\examples{
MakeFiles(VarTranslation, base_path, param_path, xmls_path)

}
