% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateNewParameterFiles.R
\name{modifyFile}
\alias{modifyFile}
\title{Modify a parameter file}
\usage{
modifyFile(paramFile, xml1, newname, variable_names = rsortie::VariableNames)
}
\arguments{
\item{paramFile}{\code{\link[=character]{character()}} Parameter file with new values}

\item{xml1}{\code{\link[=character]{character()}} Base XML parameter file to be modified}

\item{newname}{name of the new parameter file being created. Automatically generated in makeFiles function}

\item{variable_names}{}
}
\value{

}
\description{
\code{modifyFile()} replaces the base parameter file with new, updated values.
}
\details{
This function is called by \code{makeFiles()}. Doing a value replacement (making a new parameter file)
does not require \code{modifyFile()} called explicitly.
}
\examples{
modifyFile(newsortievals,xml1, newname)

}
