% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParseXML.R
\name{parseMap}
\alias{parseMap}
\title{Parse spatial files}
\usage{
parseMap(grid_data)
}
\arguments{
\item{grid_data}{\code{\link[=character]{character()}} Grid data}
}
\value{
The file sprd.
}
\description{
\code{parseMap()} takes the output XML file name and returns the file sprd.
}
\examples{
parseMap(grid_data)

}
