% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParseXML.R
\name{parseMap}
\alias{parseMap}
\title{Parse spatial files}
\usage{
parseMap(xmlname)
}
\arguments{
\item{xmlname}{\code{\link[=character]{character()}} XML file name}
}
\value{
The output contained in grids from SORTIE run as a data frame
}
\description{
\code{parseMap()} takes the output XML file name and returns the outputs stored in maps.
}
\examples{
parseMap(grid_data)

}
