% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParseXML.R
\name{ParseMap}
\alias{ParseMap}
\title{Parse spatial files}
\usage{
ParseMap(grid_data)
}
\arguments{
\item{grid_data}{\code{\link[=character]{character()}} Grid data}
}
\value{
The file sprd.
}
\description{
\code{ParseMap()} takes the output XML file name and returns the file sprd.
}
\examples{
ParseMap(grid_data)

}
