% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParseXML.R
\name{parseXML}
\alias{parseXML}
\title{Parse output data frame}
\usage{
parseXML(xmlname)
}
\arguments{
\item{xmlname}{\code{\link[=character]{character()}} XML file name}
}
\value{
The file output data frame.
}
\description{
\code{parseXML()} takes the SORTIE output XML file name and returns the file output data frame
}
\examples{
parseXML(xmlname)

}
