% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParseXML.R
\name{ParseXML}
\alias{ParseXML}
\title{Parse output data frame}
\usage{
ParseXML(xmlname)
}
\arguments{
\item{xmlname}{\code{\link[=character]{character()}} XML file name}
}
\value{
The file output data frame.
}
\description{
\code{ParseXML()} takes the SORTIE output XML file name and returns the file output data frame
}
\examples{
ParseXML(xmlname)

}
