% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateNewParameterFiles.R
\name{prepareFile}
\alias{prepareFile}
\title{Prepare a file to be updated}
\usage{
prepareFile(pfname)
}
\arguments{
\item{pfname}{\code{\link[=character]{character()}} File path and name of the parameter file with new values}
}
\value{
pf1
}
\description{
\code{prepareFile()} prepares the new parameter file for \code{modifyFile()}
}
\examples{
prepareFile(paramFile)

}
