% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateNewParameterFiles.R
\name{prepareFile}
\alias{prepareFile}
\title{Match formatting of the new values file to the base SORTIE parameter file}
\usage{
prepareFile(pfname)
}
\arguments{
\item{pfname}{\code{\link[=character]{character()}} File path and name of the parameter file with new values}
}
\value{
a formatted table containing the values of
}
\description{
\code{prepareFile()} is called by \code{modifyFile()} and receives an table of values that will replace the
values already in the base SORTIE parameter file. Typically this function is not called outside of the
\code{makeFiles()} function sequence.
}
\examples{
prepareFile(paramFile)

}
