% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{prepInputs}
\alias{prepInputs}
\title{Prepare inputs}
\usage{
prepInputs(fileNames, varTranslation)
}
\arguments{
\item{fileNames}{The text file name that contains all the file names to update and be
updated.
This file must be a csv with a type and}

\item{varTranslation}{The translation file that recognizes the behaviour and parameter names in the base
parameter file and translates given names of objects from R to ensure the right parameter value is updated

There are many different types of parameter values that can be updated.
The \code{\link[=varTranslation]{varTranslation()}} file is then essential to ensure the parameter name that is defined in the R environment
can be found in the base parameter file to replace the correct value. This csv file must be setup correctly
with the following columns:#'
col 1: input parameter name defined
col 2: type
col 3: name in the line being replaced in the base parameter file
col 4: group name

Column 2 defines the type of parameter to be replaced and these are the valid values:
1 = basic case: variable parameter is directly after the name
2 = basic case with species: same, but with a species name after it
3 = behaviorlist type: but basic parameter - similar to 1
4 = behaviorlist type: but with species - similar to 2
5 = output files: so parameter file will have a directory name
6 = groups with species on the previous line, e.g. for initial density}
}
\value{

}
\description{
\code{prepInputs()} generates the base parameter values file to be modified
}
\examples{
prepInputs(fileNames, varTranslation)

}
