% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{read_keyhole}
\alias{read_keyhole}
\title{Read in spatial files}
\usage{
read_keyhole(file)
}
\arguments{
\item{file}{\code{\link[=character]{character()}} KMZ File}
}
\value{

}
\description{
\code{read_keyhole()} is a helper function to read in .kmz files
}
\examples{
read_keyhole(file)

}
