% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{read_keyhole}
\alias{read_keyhole}
\title{Read Keyhole}
\usage{
read_keyhole(file)
}
\arguments{
\item{file}{\code{\link[=character]{character()}} KMZ File}
}
\value{

}
\description{
Read Keyhole
}
\examples{
read_keyhole(file)

}
