% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutputFunctions.R
\name{readPlotFile}
\alias{readPlotFile}
\title{Read plot file}
\usage{
readPlotFile(outdir)
}
\arguments{
\item{outdir}{\code{\link[=character]{character()}} Output directory}
}
\value{

}
\description{
\code{readPlotFile()} reads .out output files.
}
\examples{
readPlotFile(outdir)

}
