% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{ReadPlotFile}
\alias{ReadPlotFile}
\title{ReadPlotFile}
\usage{
ReadPlotFile(outdir)
}
\arguments{
\item{outdir}{\code{\link[=character]{character()}} Output directory}
}
\value{

}
\description{
ReadPlotFile
}
\examples{
ReadPlotFile(outdir)

}
