% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReplaceFunctions.R
\name{removeRow}
\alias{removeRow}
\title{Remove a row}
\usage{
removeRow(ln1, rf)
}
\arguments{
\item{ln1}{\code{\link[=double]{double()}} Row number to remove}

\item{rf}{\code{\link[=character]{character()}} Base XML file to be modified}
}
\description{
\code{removeRow()} removes a row from the base parameter file.
}
\examples{
removeRow(lnm,xml1)

}
