% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReplaceFunctions.R
\name{removeSpecies}
\alias{removeSpecies}
\title{Remove a species}
\usage{
removeSpecies(sp, rf)
}
\arguments{
\item{sp}{\code{\link[=character]{character()}} Species name}

\item{rf}{\code{\link[=character]{character()}} Base XML file to be modified}
}
\description{
\code{removeSpecies()} removes a species from the base parameter file.
}
\examples{
RemoveSpecies(sp, rf)

}
