% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{RemoveSpecies}
\alias{RemoveSpecies}
\title{Remove a species}
\usage{
RemoveSpecies(sp, rf)
}
\arguments{
\item{sp}{\code{\link[=character]{character()}} Species name}

\item{rf}{\code{\link[=character]{character()}} Base XML file to be modified}
}
\value{

}
\description{
\code{RemoveSpecies()} removes a species from the base parameter file and simulation.
}
\examples{
RemoveSpecies(sp, rf)

}
