% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReplaceFunctions.R
\name{replaceInfo}
\alias{replaceInfo}
\title{Replace parameter values in base XML file}
\usage{
replaceInfo(rf, variable_names, pf1, ncols, newname)
}
\arguments{
\item{rf}{\code{\link[=character]{character()}} Base XML parameter file to be modified}

\item{variable_names}{\code{\link[=list]{list()}} Variable Names translation file}

\item{pf1}{\code{\link[=character]{character()}} Parameter file with new values}

\item{ncols}{\code{\link[=double]{double()}} Number of columns}

\item{newname}{\code{\link[=character]{character()}} New file name}
}
\value{

}
\description{
\code{replaceInfo()} updates the base parameter file with the new values from the new parameter values file.
}
\examples{
xml2 <- ReplaceInfo(xml1, variable_names, pf1, ncols, newname)

}
