% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReplaceInfo.R
\name{ReplaceLines}
\alias{ReplaceLines}
\title{Replace Lines}
\usage{
ReplaceLines(rf, pf1)
}
\arguments{
\item{rf}{\code{\link[=character]{character()}} Base XML file to be modified}

\item{pf1}{\code{\link[=character]{character()}} Parameter file with new values}
}
\value{

}
\description{
\code{ReplaceLines()} replaces a chunk of the parameter file with a different set of lines. It assumes the new set of lines are complete and completely replace relevant section.
}
\examples{
xml2 <- ReplaceLines(xml1, pf1)

}
