% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{ReplaceParameter}
\alias{ReplaceParameter}
\title{ReplaceParameter}
\usage{
ReplaceParameter(ln1, rf, varvalue)
}
\arguments{
\item{ln1}{\code{\link[=double]{double()}} Line number of parameter to replace}

\item{rf}{\code{\link[=character]{character()}} Base XML parameter file to be modified}

\item{varvalue}{\code{\link[=character]{character()}} Output file name}
}
\value{

}
\description{
ReplaceParameter
}
\examples{
ReplaceParameter(ln1, rf, fname)

}
