% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{RunSortie}
\alias{RunSortie}
\title{Run the SORTIE model}
\usage{
RunSortie(fname, sortie_loc)
}
\arguments{
\item{fname}{\code{\link[=character]{character()}} File path and name to be run}

\item{sortie_loc}{SORTIE location '0'}
}
\value{

}
\description{
\code{RunSortie()} is a wrapper function that takes the modified parameter file with new values and runs the simulation through the SORTIE-ND GUI.
}
\examples{
RunSortie(paste0(xmls_path,Xmls2Run[ix]),0)

}
