% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunSortieFunctions.R
\name{runSortie}
\alias{runSortie}
\title{Run the SORTIE model}
\usage{
runSortie(fname, sortie_loc = 0)
}
\arguments{
\item{fname}{\code{\link[=character]{character()}} File path and name of SORTIE parameter file to be run}

\item{sortie_loc}{Location of SORTIE program}
}
\description{
\code{runSortie()} passes a parameter file to SORTIE to start a simulation.
}
\details{
default location for sortie_loc = 0: "C:/Program Files (x86)/SORTIE/bin/coremodel.exe/",
if the program is installed else, use sortie_loc = 1 and specify the location of coremodel.exe,
end the character string of the core model location after coremodel.exe (i.e don't add a
final forward slash)
}
\examples{
# MySortieLoc <- "C:/Program Files/SORTIE/bin/coremodel.exe"
#(runSortie(samplebasexml,0))

}
