% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunSortieFunctions.R
\name{runSortie}
\alias{runSortie}
\title{Run the SORTIE model}
\usage{
runSortie(fname, sortie_loc)
}
\arguments{
\item{fname}{\code{\link[=character]{character()}} File path and name to be run}

\item{sortie_loc}{SORTIE location '0'}
}
\value{

}
\description{
\code{runSortie()} is a wrapper function that passes a parameter file to SORTIE to start a simulation.
}
\examples{
#(runSortie(parameterfile.xml,0))

}
