% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateNewParameterFiles.R
\name{treelistDfn}
\alias{treelistDfn}
\title{Add new variables that translate updates to base SORTIE parameter file}
\usage{
treelistDfn(initname, numDigits = 0, diamMin, diamMax, diamInc)
}
\arguments{
\item{initname}{Naming convention \code{\link[=character]{character()}} used to describe diameter size classes}

\item{numDigits}{Number of decimal places required to match the base parameter file}

\item{diamMin}{Minimum diameter size}

\item{diamMax}{Maximum diameter size}

\item{diamInc}{Size (in cm) of diameter bins}
}
\value{

}
\description{
\code{treelistDfn()} adds additional initial tree diameter size classes and prefixes
to \code{VariableNames}
}
\details{
\code{VariableNames} is a table that translates the names of parameters found within behaviours
(variables) defined by a user in the newvals object to the names of these parameters (variables) found in
the base SORTIE parameter file. This file is essential to finding the right variable within the right
behaviour to update with new values during a \code{makeFiles()} call.

For details on how to write a new \code{VariableNames} file, see the \emph{Structure of rsortie} vignette, linked below

There is a default \code{VariableNames} loaded with the rsortie package, but a user may wish to add additional
variable translations.
}
\examples{
samplebasexml[1:30]
VariableNames[1:30,]
}
\seealso{
\href{https://aclason.github.io/rsortie/articles/use_rsortie.html}{Structure of rsortie}
}
