% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReplaceFunctions.R
\name{updateNumYears}
\alias{updateNumYears}
\title{Update Number of Years (timesteps)}
\usage{
updateNumYears(xmls, num_years)
}
\arguments{
\item{xmls}{\code{\link[=character]{character()}} XML file to be modified with pathway included}

\item{num_years}{\code{\link[=numeric]{numeric()}} Number of years to run a simulation}
}
\description{
\code{updateNumYears()} Updates the number of years for a simulation
}
\examples{
updateNumYears(xmls, num_years)

}
